feat: add some design features
