+++ .eslintignore
@@ -1,1 +1,1 @@
-/src/components/shared/figures.img.tsx
+/src/components/shared/figure/figures.img.tsx
