+++ src/components/header/replay-header/replay-header.tsx
@@ -3,7 +3,7 @@ import { FigureColor } from '../../../enums/enums';
 import { RootState } from '../../../redux/reducers';
 import { setReplaySpeed } from '../../../redux/reducers/replays.state';
 import { intervals } from '../../../utils/usefull_utils';
-import ButtonPopup from '../../shared/button.styled.component';
+import ButtonPopup from '../../shared/button-styled/button.styled.component';
 
 function ReplayHeader() {
   const interval = useSelector((state:RootState) => state.replays.speed);
