+++ src/components/main/menu/menu.style.scss
@@ -28,16 +28,41 @@
       width: 76px;
       height: 76px;
       cursor: pointer;
-      transition: background-color .2s;
     }
     .button-online{
       right: 0;
       top: 0;
+      transition: all .2s;
     }
     .button-replays{
       left: 0;
       bottom: 0;
     }
   }
-  
+  button:hover{
+    animation: shake 1.5s;
+  }
+
+
+}
+
+@keyframes shake {
+  0% {
+    transform: translate(1px, 1px) rotate(0deg);
+  }
+  10% {
+    transform: translate(-1px, -2px) rotate(-1deg);
+  }
+  30% {
+    transform: translate(3px, 2px) rotate(0deg);
+  }
+  50% {
+    transform: translate(-1px, 2px) rotate(-1deg);
+  }
+  70% {
+    transform: translate(3px, 1px) rotate(-1deg);
+  }
+  90% {
+    transform: translate(1px, 2px) rotate(0deg);
+  }
 }
