+++ src/components/popup/replays/replay.component.tsx
@@ -2,7 +2,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { FigureColor, GridColor } from '../../../enums/enums';
 import { IReplayRes } from '../../../interfaces/interfaces';
 import { timeFormatHandle } from '../../../utils/timer.string-maker';
-import ButtonPopup from '../../shared/button.styled.component';
+import ButtonPopup from '../../shared/button-styled/button.styled.component';
 import winnerIMG from '../../../assets/winner-icon.png';
 import {
   endGameButton,
