+++ src/components/shared/player-stats/player.stats.scss
@@ -12,12 +12,14 @@ $table-width: 136px;
     justify-content: center;
     width: $header-size;
     height: $header-size;
+    color:white;
     border: 5px solid #e8a87c;
     border-radius: 50%;
     text-align: center;
-    color: #40b3a2;
+    background-color: grey;
     margin-bottom: 21px;
     position: relative;
+    transition: background-color .5s;
     .player-mover {
       display: flex;
       align-items:center;
@@ -44,12 +46,13 @@ $table-width: 136px;
     padding: 4px 0;
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
     border: 5px solid #f0f7f4;
-    background: #40b3a2;
+    background: grey;
     color: #f0f7f4;
     font-family: Open Sans;
     text-align: center;
     font-size: 24px;
     margin-bottom: 21px;
+    transition: background-color .5s;
     cursor:default;
   }
